centralized dependencies from inttest job build files
