+++ inttest/build.gradle
@@ -1,13 +1,65 @@
-plugins {
-    id 'java'
+ext {
+    jarDir = "$projectDir/../build/libs"
 }
-
 apply from: '../dependencies.gradle'
 
-repositories {
-    mavenCentral()
+allprojects {
+    apply plugin: 'java'
+
+    repositories {
+        mavenCentral()
+    }
+
+    dependencies {
+        testCompile junitDependency
+    }
 }
 
-dependencies {
-    testCompile junitDependency
+subprojects {
+
+    buildscript {
+        ext {
+            //dependencies
+            postgresqlVersion = '42.1.4'
+            // TODO no snapshot versions
+            ioxIliVersion = '1.20.0-SNAPSHOT'
+            ilivalidatorVersion = '1.6.0-SNAPSHOT'
+            ili2pgVersion = '3.11.0-SNAPSHOT'
+            ioxWkfVersion = '0.9.0-SNAPSHOT'
+            sqliteJdbcVersion = '3.8.11.2'
+            derbyVersion = '10.13.1.1'
+        }
+
+        repositories {
+            maven {
+                url "http://download.osgeo.org/webdav/geotools/"
+            }
+            mavenCentral()
+            maven {
+                url "http://jars.interlis.ch"
+            }
+            maven {
+                url "http://jars.umleditor.org"
+            }
+        }
+
+        configurations.all {
+            resolutionStrategy {
+                force group: 'ch.interlis', name: 'iox-ili', version: ioxIliVersion
+            }
+        }
+
+        dependencies {
+            classpath fileTree(dir: rootProject.ext.jarDir, include: '*.jar')
+
+            //compile group: 'org.postgresql', name: 'postgresql', version: postgresqlVersion
+
+            classpath group: 'ch.interlis', name: 'ilivalidator', version: ilivalidatorVersion
+            classpath group: 'ch.interlis', name: 'ili2pg', version: ili2pgVersion
+            classpath group: 'ch.interlis', name: 'iox-wkf', version: ioxWkfVersion
+            classpath group: 'ch.interlis', name: 'iox-ili', version: ioxIliVersion
+        }
+    }
 }
+
+
