+++ inttest/jobs/iliValidator/build.gradle
@@ -1,59 +1,5 @@
 import ch.so.agi.gretl.tasks.*
 
-apply plugin: 'ch.so.agi.gretl'
-
-buildscript {
-    ext {
-        //dependencies
-        postgresqlVersion = '42.1.4'
-        // TODO no snapshot versions
-        ioxIliVersion = '1.20.0-SNAPSHOT'
-        ilivalidatorVersion = '1.6.0-SNAPSHOT'
-        ili2pgVersion = '3.11.0-SNAPSHOT'
-        ioxWkfVersion = '0.9.0-SNAPSHOT'
-        sqliteJdbcVersion = '3.8.11.2'
-        derbyVersion = '10.13.1.1'
-    }
-
-
-    repositories {
-        maven {
-            url "http://download.osgeo.org/webdav/geotools/"
-        }
-        mavenCentral()
-        maven {
-            url "http://jars.interlis.ch"
-        }
-        maven {
-            url "http://jars.umleditor.org"
-        }
-        flatDir {
-            dirs '../../../build/libs'
-        }
-    }
-
-    configurations.all {
-        resolutionStrategy {
-            force group: 'ch.interlis', name: 'iox-ili', version: '1.20.0-SNAPSHOT'
-        }
-    }
-
-    dependencies {
-        classpath "ch.so.agi.gretl:gretl:1.0.4-SNAPSHOT"
-
-        //compile group: 'org.postgresql', name: 'postgresql', version: postgresqlVersion
-
-        classpath group: 'ch.interlis', name: 'ilivalidator', version: ilivalidatorVersion
-        classpath group: 'ch.interlis', name: 'ili2pg', version: ili2pgVersion
-        classpath group: 'ch.interlis', name: 'iox-wkf', version: ioxWkfVersion
-        classpath group: 'ch.interlis', name: 'iox-ili', version: ioxIliVersion
-
-        //compile group: 'org.xerial', name: 'sqlite-jdbc', version: sqliteJdbcVersion
-        //compile group: 'org.apache.derby', name: 'derby', version: derbyVersion
-    }
-}
-
-
 defaultTasks 'validate'
 
 task validate(type: IliValidator){
