+++ inttest/src/test/java/ch/so/agi/gretl/util/TestUtil.java
@@ -1,0 +1,35 @@
+package ch.so.agi.gretl.util;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+
+public class TestUtil {
+
+    public static void runJob(String jobPath) throws Exception {
+        Process p = Runtime.getRuntime().exec("./gradlew --project-dir " + jobPath);
+
+        BufferedReader stdInput = new BufferedReader(new InputStreamReader(p.getInputStream()));
+        BufferedReader stdError = new BufferedReader(new InputStreamReader(p.getErrorStream()));
+
+        String s;
+
+        // read the output from the command
+        System.out.println("Here is the standard output of the command:\n");
+        while ((s = stdInput.readLine()) != null) {
+            System.out.println(s);
+        }
+
+        // read any errors from the attempted command
+        System.out.println("Here is the standard error of the command (if any):\n");
+        while ((s = stdError.readLine()) != null) {
+            System.out.println(s);
+        }
+
+        p.waitFor();
+
+        assertThat(p.exitValue(), is(0));
+    }
+}
